leetcode/0257.Binary-Tree-Paths/257.BinaryTreePaths.go
package leetcode

import (
	"strconv"
	"strings"
)

type TreeNode struct {
	Val   int
	Left  *TreeNode
	Right *TreeNode
}

func binaryTreePaths(root *TreeNode) []string {
	var ans []string
	var dfs func(node *TreeNode, path []string)
	dfs = func(node *TreeNode, path []string) {
		if node == nil {
			return
		}
		path = append(path, strconv.Itoa(node.Val))
		defer func() { path = path[:len(path)-1] }()
		if node.Left == nil && node.Right == nil {
			ans = append(ans, strings.Join(path, "->"))
			return
		}
		dfs(node.Left, path)
		dfs(node.Right, path)
	}
	dfs(root, []string{})
	return ans
}
